Fixes to Shipwright Build Strategy (#67)
* fix: Use UBI Images for Bash

Use the ubi-minimal image for running pure bash commands. This avoids
pulling additional images from Docker Hub, which requires a pull secret
to be configured.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>

* fix: Use Overridable Volumes in Shipwright Build Strategy

Use overridable volumes in the cluster build strategy for buildpacks.
This ensures the build strategy works "out of the box", and does not
require additional configuration from developers to execute.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>

---------

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>